moved cell types. fixes #353
