+++ bsb/cell_types.py
@@ -2,12 +2,12 @@
     Module for the CellType configuration node and its dependencies.
 """
 
-from .. import config
-from ..config import types
-from ..placement import PlacementStrategy
-from ..placement.indicator import PlacementIndications
-from ..helpers import SortableByAfter
-from ..exceptions import *
+from . import config
+from .config import types
+from .placement import PlacementStrategy
+from .placement.indicator import PlacementIndications
+from .helpers import SortableByAfter
+from .exceptions import *
 import abc
