--- bsb/objects/__init__.py
@@ -1,1 +1,0 @@
-from .cell_type import CellType
