--- docs/bsb/bsb.objects.rst
@@ -1,21 +1,0 @@
-bsb.objects package
-===================
-
-Submodules
-----------
-
-bsb.objects.cell\_type module
------------------------------
-
-.. automodule:: bsb.objects.cell_type
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: bsb.objects
-   :members:
-   :undoc-members:
-   :show-inheritance:
